Fix: Moved student routes into their own folder.
