+++ routes/student/deleteStudent.js
@@ -1,0 +1,6 @@
+const express = require('express');
+const router = express.Router();
+const studentController = require('../controllers/student');
+const { ensureAuth } = require('../middleware/auth');
+
+router.delete('/deleteStudent', ensureAuth, studentController.deleteStudent);
