--- routes/teacher.js
@@ -1,8 +1,0 @@
-// const express = require('express');
-// const router = express.Router();
-// const teacherController = require('../controllers/teacher');
-// const { ensureAuth } = require('../middleware/auth');
-
-// router.get('/studentList', ensureAuth, teacherController.showStudentList);
-
-// module.exports = router;
